addition of NVisII and iGibson renderers

+++ robosuite/environments/base.py
@@ -2,7 +2,9 @@ from collections import OrderedDict
 from mujoco_py import MjSim, MjRenderContextOffscreen
 from mujoco_py import load_model_from_xml
 
-from robosuite.utils import SimulationError, XMLError, MujocoPyRenderer
+from robosuite.utils import SimulationError, XMLError
+from robosuite.renderers.mujoco.mujoco_py_renderer import MujocoPyRenderer
+from robosuite.renderers.base import load_renderer_config
 import robosuite.utils.macros as macros
 from robosuite.models.base import MujocoModel
 
@@ -18,17 +20,13 @@ def register_env(target_class):
 def make(env_name, *args, **kwargs):
     """
     Instantiates a robosuite environment.
-
     This method attempts to mirror the equivalent functionality of gym.make in a somewhat sloppy way.
-
     Args:
         env_name (str): Name of the robosuite environment to initialize
         *args: Additional arguments to pass to the specific environment class initializer
         **kwargs: Additional arguments to pass to the specific environment class initializer
-
     Returns:
         MujocoEnv: Desired robosuite environment
-
     Raises:
         Exception: [Invalid environment name]
     """
@@ -58,38 +56,29 @@ class EnvMeta(type):
 class MujocoEnv(metaclass=EnvMeta):
     """
     Initializes a Mujoco Environment.
-
     Args:
         has_renderer (bool): If true, render the simulation state in
             a viewer instead of headless mode.
-
         has_offscreen_renderer (bool): True if using off-screen rendering.
-
         render_camera (str): Name of camera to render if `has_renderer` is True. Setting this value to 'None'
             will result in the default angle being applied, which is useful as it can be dragged / panned by
             the user using the mouse
-
         render_collision_mesh (bool): True if rendering collision meshes
             in camera. False otherwise.
-
         render_visual_mesh (bool): True if rendering visual meshes
             in camera. False otherwise.
-
         render_gpu_device_id (int): corresponds to the GPU device id to use for offscreen rendering.
             Defaults to -1, in which case the device will be inferred from environment variables
             (GPUS or CUDA_VISIBLE_DEVICES).
-
         control_freq (float): how many control signals to receive
             in every simulated second. This sets the amount of simulation time
             that passes between every action input.
-
         horizon (int): Every episode lasts for exactly @horizon timesteps.
-
         ignore_done (bool): True if never terminating the environment (ignore @horizon).
-
         hard_reset (bool): If True, re-loads model, sim, and render object upon a reset call, else,
             only calls sim.reset and resets all robosuite-internal variables
-
+        renderer (str): string for the renderer to use
+        renderer_config (dict): dictionary for the renderer configurations
     Raises:
         ValueError: [Invalid renderer selection]
     """
@@ -105,7 +94,9 @@ class MujocoEnv(metaclass=EnvMeta):
         control_freq=20,
         horizon=1000,
         ignore_done=False,
-        hard_reset=True
+        hard_reset=True,
+        renderer="mujoco",
+        renderer_config=None,
     ):
         # First, verify that both the on- and off-screen renderers are not being used simultaneously
         if has_renderer is True and has_offscreen_renderer is True:
@@ -134,6 +125,9 @@ class MujocoEnv(metaclass=EnvMeta):
         self.control_timestep = None
         self.deterministic_reset = False            # Whether to add randomized resetting of objects / robot joints
 
+        self.renderer = renderer
+        self.renderer_config = renderer_config
+
         # Load the model
         self._load_model()
 
@@ -143,16 +137,46 @@ class MujocoEnv(metaclass=EnvMeta):
         # Initialize the simulation
         self._initialize_sim()
 
+        #initializes the rendering
+        self.initialize_renderer()
+
         # Run all further internal (re-)initialization required
         self._reset_internal()
 
         # Load observables
-        self._observables = self._setup_observables()
+        if hasattr(self.viewer, '_setup_observables'):
+            self._observables = self.viewer._setup_observables()
+        else:
+            self._observables = self._setup_observables()
+
+        # check if viewer has get observations method and set a flag for future use.
+        self.viewer_get_obs = hasattr(self.viewer, '_get_observations')
+
+    def initialize_renderer(self):
+        self.renderer = self.renderer.lower()
+
+        if self.renderer_config is None and self.renderer != 'mujoco':
+            self.renderer_config = load_renderer_config(self.renderer)
+
+        if self.renderer == 'mujoco' or self.renderer == 'default':
+            pass
+        elif self.renderer == 'nvisii':
+            from robosuite.renderers.nvisii.nvisii_renderer import NVISIIRenderer
+
+            self.viewer = NVISIIRenderer(env=self,
+                                         **self.renderer_config)
+        elif self.renderer == 'igibson':
+            from robosuite.renderers.igibson.igibson_renderer import iGibsonRenderer
+
+            self.viewer = iGibsonRenderer(env=self,
+                                         **self.renderer_config
+                                         )
+        else:
+            raise ValueError(f'{self.renderer} is not a valid renderer name. Valid options include default (native mujoco renderer), nvisii, and igibson')
 
     def initialize_time(self, control_freq):
         """
         Initializes the time constants used for simulation.
-
         Args:
             control_freq (float): Hz rate to run control loop at within the simulation
         """
@@ -168,7 +192,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def set_model_postprocessor(self, postprocessor):
         """
         Sets the post-processor function that self.model will be passed to after load_model() is called during resets.
-
         Args:
             postprocessor (None or function): If set, postprocessing method should take in a Task-based instance and
                 return no arguments.
@@ -199,7 +222,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def _setup_observables(self):
         """
         Sets up observables to be used for this environment.
-
         Returns:
             OrderedDict: Dictionary mapping observable names to its corresponding Observable object
         """
@@ -209,7 +231,6 @@ class MujocoEnv(metaclass=EnvMeta):
         """
         Creates a MjSim object and stores it in self.sim. If @xml_string is specified, the MjSim object will be created
         from the specified xml_string. Else, it will pull from self.model to instantiate the simulation
-
         Args:
             xml_string (str): If specified, creates MjSim object from this filepath
         """
@@ -226,20 +247,22 @@ class MujocoEnv(metaclass=EnvMeta):
     def reset(self):
         """
         Resets simulation.
-
         Returns:
             OrderedDict: Environment observation space after reset occurs
         """
         # TODO(yukez): investigate black screen of death
         # Use hard reset if requested
+
         if self.hard_reset and not self.deterministic_reset:
-            self._destroy_viewer()
+            if self.renderer == 'mujoco' or self.renderer == 'default':
+                self._destroy_viewer()
             self._load_model()
             self._postprocess_model()
             self._initialize_sim()
         # Else, we only reset the sim internally
         else:
             self.sim.reset()
+        
         # Reset necessary robosuite-centric variables
         self._reset_internal()
         self.sim.forward()
@@ -252,8 +275,14 @@ class MujocoEnv(metaclass=EnvMeta):
                 self.modify_observable(observable_name=obs_name, attribute="sensor", modifier=obs._sensor)
         # Make sure that all sites are toggled OFF by default
         self.visualize(vis_settings={vis: False for vis in self._visualizations})
+
+        if self.viewer is not None and self.renderer != 'mujoco':
+            self.viewer.reset()
+        
+        observations = self.viewer._get_observations(force_update=True) if self.viewer_get_obs else self._get_observations(force_update=True)
+
         # Return new observations
-        return self._get_observations(force_update=True)
+        return observations
 
     def _reset_internal(self):
         """Resets simulation internal configurations."""
@@ -297,7 +326,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def _update_observables(self, force=False):
         """
         Updates all observables in this environment
-
         Args:
             force (bool): If True, will force all the observables to update their internal values to the newest
                 value. This is useful if, e.g., you want to grab observations when directly setting simulation states
@@ -309,15 +337,12 @@ class MujocoEnv(metaclass=EnvMeta):
     def _get_observations(self, force_update=False):
         """
         Grabs observations from the environment.
-
         Args:
             force_update (bool): If True, will force all the observables to update their internal values to the newest
                 value. This is useful if, e.g., you want to grab observations when directly setting simulation states
                 without actually stepping the simulation.
-
         Returns:
             OrderedDict: OrderedDict containing observations [(name_string, np.array), ...]
-
         """
         observations = OrderedDict()
         obs_by_modality = OrderedDict()
@@ -350,21 +375,16 @@ class MujocoEnv(metaclass=EnvMeta):
     def step(self, action):
         """
         Takes a step in simulation with control command @action.
-
         Args:
             action (np.array): Action to execute within the environment
-
         Returns:
             4-tuple:
-
                 - (OrderedDict) observations from the environment
                 - (float) reward from the environment
                 - (bool) whether the current episode is completed or not
                 - (dict) misc information
-
         Raises:
             ValueError: [Steps past episode termination]
-
         """
         if self.done:
             raise ValueError("executing action in terminated episode")
@@ -390,12 +410,16 @@ class MujocoEnv(metaclass=EnvMeta):
         self.cur_time += self.control_timestep
 
         reward, done, info = self._post_action(action)
-        return self._get_observations(), reward, done, info
+
+        if self.viewer is not None and self.renderer != 'mujoco':
+            self.viewer.update()
+
+        observations = self.viewer._get_observations() if self.viewer_get_obs else self._get_observations()
+        return observations, reward, done, info
 
     def _pre_action(self, action, policy_step=False):
         """
         Do any preprocessing before taking an action.
-
         Args:
             action (np.array): Action to execute within the environment
             policy_step (bool): Whether this current loop is an actual policy step or internal sim update step
@@ -405,17 +429,13 @@ class MujocoEnv(metaclass=EnvMeta):
     def _post_action(self, action):
         """
         Do any housekeeping after taking an action.
-
         Args:
             action (np.array): Action to execute within the environment
-
         Returns:
             3-tuple:
-
                 - (float) reward from the environment
                 - (bool) whether the current episode is completed or not
                 - (dict) empty dict to be filled with information by subclassed method
-
         """
         reward = self.reward(action)
 
@@ -427,10 +447,8 @@ class MujocoEnv(metaclass=EnvMeta):
     def reward(self, action):
         """
         Reward should be a function of state and action
-
         Args:
             action (np.array): Action to execute within the environment
-
         Returns:
             float: Reward from environment
         """
@@ -442,19 +460,29 @@ class MujocoEnv(metaclass=EnvMeta):
         """
         self.viewer.render()
 
+    def get_pixel_obs(self):
+        """
+        Gets the pixel observations for the environment from the specified renderer
+        """
+        self.viewer.get_pixel_obs()
+
+    def close_renderer(self):
+        """
+        Closes the renderer
+        """
+        self.viewer.close()
+
     def observation_spec(self):
         """
         Returns an observation as observation specification.
-
         An alternative design is to return an OrderedDict where the keys
         are the observation names and the values are the shapes of observations.
         We leave this alternative implementation commented out, as we find the
         current design is easier to use in practice.
-
         Returns:
             OrderedDict: Observations from the environment
         """
-        observation = self._get_observations()
+        observation = self.viewer._get_observations() if self.viewer_get_obs else self._get_observations()
         return observation
 
     def clear_objects(self, object_names):
@@ -462,7 +490,6 @@ class MujocoEnv(metaclass=EnvMeta):
         Clears objects with the name @object_names out of the task space. This is useful
         for supporting task modes with single types of objects, as in
         @self.single_object_mode without changing the model definition.
-
         Args:
             object_names (str or list of str): Name of object(s) to remove from the task workspace
         """
@@ -474,7 +501,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def visualize(self, vis_settings):
         """
         Do any needed visualization here
-
         Args:
             vis_settings (dict): Visualization keywords mapped to T/F, determining whether that specific
                 component should be visualized. Should have "env" keyword as well as any other relevant
@@ -484,16 +510,22 @@ class MujocoEnv(metaclass=EnvMeta):
         for obj in self.model.mujoco_objects:
             obj.set_sites_visibility(sim=self.sim, visible=vis_settings["env"])
 
+    def set_camera_pos_quat(self, camera_pos, camera_quat):
+        if self.renderer == "nvisii":
+            self.viewer.set_camera_pos_quat(camera_pos, camera_quat)
+        else:
+            print('Setting camera position and quat requires NVISII renderer.')
+
     def reset_from_xml_string(self, xml_string):
         """
         Reloads the environment from an XML description of the environment.
-
         Args:
             xml_string (str): Filepath to the xml file that will be loaded directly into the sim
         """
 
         # if there is an active viewer window, destroy it
-        self.close()
+        if self.renderer != 'nvisii':
+            self.close()
 
         # Since we are reloading from an xml_string, we are deterministically resetting
         self.deterministic_reset = True
@@ -510,14 +542,12 @@ class MujocoEnv(metaclass=EnvMeta):
     def check_contact(self, geoms_1, geoms_2=None):
         """
         Finds contact between two geom groups.
-
         Args:
             geoms_1 (str or list of str or MujocoModel): an individual geom name or list of geom names or a model. If
                 a MujocoModel is specified, the geoms checked will be its contact_geoms
             geoms_2 (str or list of str or MujocoModel or None): another individual geom name or list of geom names.
                 If a MujocoModel is specified, the geoms checked will be its contact_geoms. If None, will check
                 any collision with @geoms_1 to any other geom in the environment
-
         Returns:
             bool: True if any geom in @geoms_1 is in contact with any geom in @geoms_2.
         """
@@ -545,13 +575,10 @@ class MujocoEnv(metaclass=EnvMeta):
         """
         Checks for any contacts with @model (as defined by @model's contact_geoms) and returns the set of
         geom names currently in contact with that model (excluding the geoms that are part of the model itself).
-
         Args:
             model (MujocoModel): Model to check contacts for.
-
         Returns:
             set: Unique geoms that are actively in contact with this model.
-
         Raises:
             AssertionError: [Invalid input type]
         """
@@ -571,7 +598,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def add_observable(self, observable):
         """
         Adds an observable to this environment.
-
         Args:
             observable (Observable): Observable instance.
         """
@@ -583,7 +609,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def modify_observable(self, observable_name, attribute, modifier):
         """
         Modifies observable with associated name @observable_name, replacing the given @attribute with @modifier.
-
         Args:
              observable_name (str): Observable to modify
              attribute (str): Observable attribute to modify.
@@ -620,7 +645,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def _check_success(self):
         """
         Checks if the task has been completed. Should be implemented by subclasses
-
         Returns:
             bool: True if the task has been completed
         """
@@ -643,7 +667,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def observation_modalities(self):
         """
         Modalities for this environment's observations
-
         Returns:
             set: All observation modalities
         """
@@ -653,7 +676,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def observation_names(self):
         """
         Grabs all names for this environment's observables
-
         Returns:
             set: All observation names
         """
@@ -664,7 +686,6 @@ class MujocoEnv(metaclass=EnvMeta):
         """
         Grabs all names of enabled observables for this environment. An observable is considered enabled if its values
         are being continually computed / updated at each simulation timestep.
-
         Returns:
             set: All enabled observation names
         """
@@ -676,7 +697,6 @@ class MujocoEnv(metaclass=EnvMeta):
         Grabs all names of active observables for this environment. An observable is considered active if its value is
         being returned in the observation dict from _get_observations() call or from the step() call (assuming this
         observable is enabled).
-
         Returns:
             set: All active observation names
         """
@@ -686,7 +706,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def _visualizations(self):
         """
         Visualization keywords for this environment
-
         Returns:
             set: All components that can be individually visualized for this environment
         """
@@ -696,7 +715,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def action_spec(self):
         """
         Action specification should be implemented in subclasses.
-
         Action space is represented by a tuple of (low, high), which are two numpy
         vectors that specify the min/max action limits per dimension.
         """
@@ -706,7 +724,6 @@ class MujocoEnv(metaclass=EnvMeta):
     def action_dim(self):
         """
         Size of the action space
-
         Returns:
             int: Action space dimension
         """
